@@ -1,11 +1,11 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
 	darkMode: ['class'],
 	content: [
 		'./pages/**/*.{js,jsx}',
 		'./components/**/*.{js,jsx}',
 		'./app/**/*.{js,jsx}',
-		'./src/**/*.{js,jsx}',
+		'./lib/**/*.{js,jsx}',
 	],
 	theme: {
 		container: {
